pyproject.toml
[build-system]
requires = ["uv_build>=0.9.10,<1.0.0"]
build-backend = "uv_build"

[project]
name = "code-readiness-template"
version = "0.1.0"
description = "Codex-native FastAPI template repo for code-readiness scaffolding."
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "alembic>=1.17.1,<2.0.0",
    "fastapi>=0.118.0,<1.0.0",
    "httpx>=0.28.1,<1.0.0",
    "opentelemetry-api>=1.38.0,<2.0.0",
    "opentelemetry-exporter-otlp-proto-http>=1.38.0,<2.0.0",
    "opentelemetry-instrumentation-fastapi>=0.59b0,<1.0.0",
    "opentelemetry-sdk>=1.38.0,<2.0.0",
    "pydantic>=2.12.5,<3.0.0",
    "psycopg[binary]>=3.2.12,<4.0.0",
    "pydantic-settings>=2.11.0,<3.0.0",
    "prometheus-client>=0.23.1,<1.0.0",
    "pyyaml>=6.0.3,<7.0.0",
    "sentry-sdk[fastapi]>=2.42.1,<3.0.0",
    "sqlalchemy>=2.0.43,<3.0.0",
    "uvicorn>=0.37.0,<1.0.0",
]

[dependency-groups]
dev = [
    "deptry>=0.23.1,<1.0.0",
    "pre-commit>=4.3.0,<5.0.0",
    "pytest>=8.4.2,<9.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "pytest-randomly>=4.0.1,<5.0.0",
    "pytest-xdist>=3.8.0,<4.0.0",
    "radon>=6.0.1,<7.0.0",
    "ruff>=0.13.2,<1.0.0",
    "testcontainers[postgres]>=4.13.2,<5.0.0",
    "ty>=0.0.1a24",
    "vulture>=2.14,<3.0.0",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
src = ["src", "tests", "scripts"]
target-version = "py313"

[tool.ruff.lint]
select = ["B", "C90", "E", "F", "I", "N", "UP", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pytest.ini_options]
addopts = "--strict-config --strict-markers --cov=src/code_readiness_template --cov-report=term-missing --cov-fail-under=90 --durations=10"
markers = [
    "integration: marks tests that require real service boundaries",
]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["src/code_readiness_template"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.deptry]
known_first_party = ["code_readiness_template"]

[tool.deptry.per_rule_ignores]
DEP002 = ["psycopg", "uvicorn"]
